chore: refactor frontend
